snapshot-create-method returns specific array
If there was a snapshot created we always want to have it in the
return-value. Therefore we cannot just return the intersection-error if
there is one.
Now the create-method always returns an array with the snapshot-object
(could be NULL) and the intersection-error (could be NULL).

Attention:
I adapted the unit-tests but couldn't test them ^^. Because civix didn't
worked properly.