bugfix: Create.php rose an error on intersection-error but no snapshot
This is a general problem: There are already two intersecting snapshots.
And the user tries to create a third one.
The third one won't be created. But how should we deal with the existing
ones? Delete both? Ask the user for deleting both? And what if one of
them belongs to another user?