modified the intersection-logic
We don't check snapshots with status DONE for intersections. Because
either the contributions of those snapshots should have already an
receipt (in which case they don't appear in a snapshot at all), or the
receipt has been deleted (in which case it should be possible to
recreate them without an intersection-error).

Intersection-errors should now occur only for snapshot from different
users, because due to the revive-logic for snapshots one and the same
user could not be left with two open snapshots at all. Either he works
on with the old one, or all old snapshots will be deleted instantly.